# swaggerconfig.yaml
openapi: 3.0.0
info:
  version: 1.0.1
  title: "Code generation"
security:
  - bearerAuth: []
servers:
  - url: '{site}/{endpoint}/'
    variables:
      endpoint:
        enum:
          - api
          - apiv2
        default: api
      site:
        enum:
          - localhost
          - example.com
          - 127.0.0.1
        default: localhost
paths:
...
  /users:
    get:
      tags:
        - Users
      description: Returns a list of users
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageOffset'
      responses:
        '200':
          description: The list of users is returned from the database.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "403":
          $ref: '#/components/responses/UnAuthorized'

    post:
      tags:
        - Users
      description: creates a new users
      responses:
        '200':
          description: creates a new user based on the fields of the user after checking the inputand the permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "403":
          $ref: '#/components/responses/UnAuthorized'
  /users/{id}:
    get:
      tags:
        - Users
      description: gets a single user data
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the user
      responses:
        '200':
          description: return user data after submitting an existing id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFound'
        "403":
          $ref: '#/components/responses/UnAuthorized'

    put:
      tags:
        - Users
      description: Updates user data
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the user to get
      responses:
        '200':
          description: updates all the data of the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "405":
          $ref: '#/components/responses/InvalidInput'
        "404":
          $ref: '#/components/responses/NotFound'
        "403":
          $ref: '#/components/responses/UnAuthorized'

    delete:
      tags:
        - Users
      description: delete user data
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the user to get
      responses:
        '200':
          description: deletes all the data of the user
          content:
            application/json:
              schema:
                type: string
                example: account closed
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "405":
          $ref: '#/components/responses/InvalidInput'
        "404":
          $ref: '#/components/responses/NotFound'
        "403":
          $ref: '#/components/responses/UnAuthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum:
            - customer
            - employee
            - disabled
        name:
          type: string
        email:
          type: string
        password:
          type: string
        day_limit:
          type: integer
          example: 500
        transaction_limit:
          type: integer
...
  parameters:
    PageLimit:
      name: limit
      in: query
      description: Limits the number of items on a page
      schema:
        type: integer
    PageOffset:
      name: offset
      in: query
      description: Specifies the page number of the artists to be displayed
      schema:
        type: integer
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 401
              body:
                type: string
                example: Access token is missing
    InvalidInput:
      description: Data input is invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 405
              body:
                type: array
                items:
                  type: object
                  properties:
                    key:
                      type: string
                      example: "password"
                    value:
                      type: string
                      example: "has to be 8 characters"
    NotFound:
      description: 404 not found
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 404
              body:
                type: string
                example: Resource X was not found
    UnAuthorized:
      description: 403 not Authorized
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 403
              body:
                type: string
                example: You dont have permissions to take this action